backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic[email]==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0

# LangChain & AI
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16
langgraph==0.0.20
openai==1.10.0

# Vector Database
chromadb==0.4.22
sentence-transformers==2.3.1

# Document Processing
pypdf==3.17.4
python-docx==1.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0